Bug 2065499 - Quota manager: turn global storage usage into a signed integer, to prevent underflow to wrap to very big unsigned integer and make the storage unusable.  a=RyanVM
Original Revision: https://phabricator.services.mozilla.com/D320454

Differential Revision: https://phabricator.services.mozilla.com/D321251

+++ dom/quota/ActorsParent.cpp
@@ -8262,7 +8262,7 @@ std::pair<uint64_t, uint64_t> QuotaManager::GetUsageAndLimitForEstimate(
             // reports its own origin usage against that limit.
             if (originInfo && originInfo->LockedPersisted()) {
               return std::pair(originInfo->LockedUsage(),
-                               mTemporaryStorageLimit);
+                               static_cast<uint64_t>(mTemporaryStorageLimit));
             }
           }
 
@@ -8654,7 +8654,7 @@ QuotaManager::GetOriginInfosExceedingGlobalLimit() const {
       },
       [temporaryStorageUsage = mTemporaryStorageUsage,
        temporaryStorageLimit = mTemporaryStorageLimit,
-       doomedUsage = uint64_t{0}](const auto& originInfo) mutable {
+       doomedUsage = int64_t{0}](const auto& originInfo) mutable {
         if (temporaryStorageUsage - doomedUsage <= temporaryStorageLimit) {
           return true;
         }
